Fix missing import for sleep()
sleep() was called in views.py but it was never imported. It used to
work but was broken with fa52ba1390d9eba0f0668e0bf6e04d9cb003541e (by
yours truly) which (rightly) removed the import statement from models
but which was indirectly imported into views since it imports * from
models.