Add network model and bind instance applications to it
Add a network model, which consists of a network definition attached to a
cluster. We grant usage using Groups and introduce a "network" field to the
instance creation application, for users requiring a specific network placement
(e.g. L2 VPNs).

The cluster attribute of the instance application becomes a dynamic property,
with a getter and a setter tied to the network.

Finally, we use the network to fill in the instance's NIC parameters for the
creation request.

Signed-off-by: Apollon Oikonomopoulos <apollon@noc.grnet.gr>