Add job-bound cache-based instance locking
Use the caching layer to lock instances when actions occur. A lock is an
instance-bound key in the cache and its value is the reason the instance is
locked.

Add a new job type for the watcher, to watch specific ganeti jobs and unlock
specified instances after finishing, optionally flushing a set of cache keys as
well.

Enforce instance locking during startup, shutdown, reboot, modification and
creation.

Signed-off-by: Apollon Oikonomopoulos <apollon@noc.grnet.gr>