Replace gevent.pool.imap with its map equivalent
In gevent v1.0, the implementation of the 'gevent.pool.imap' has been
modified to return an iterator as its result. In ganetimgr, this
modification results in no data shown to the user since this iterator
is never consumed by the application.

To fix this issue, we replace the call to imap with the map equivalent,
which makes an internal call the to imap method and then consumes it
too. Finally, we also removed the call to 'gevent.pool.join' method
since it is not necessary anymore.

Conflicts:
	ganeti/models.py
	ganeti/views.py
	notifications/views.py