bug fix: get locked nodegroups and instance check
ganetimgr was trying to see if an instance is locked by getting
the nodegroup for each instance and then checking if it was locked.
Instead now it gathers all the locked nodes while descovering the locked
nodegroups and just check if an instance's node exists in the list of locked
nodes.