ApplicationReviewForm required fields and prepare for tests.
If the review has been rejected, it should not matter if a cluster
has been chosen. So these fiels should not be required in application
review form, but appear as required in case the choice is not 'reject'.

Also the aplication model (which tries to create a vm on success)
and the form, change behaviour on testing, because:

1) the database is empty and some inputs are ModelChoiceFields
2) we don't want and don't care to access ganeti rapi