[ganetimgr1.8] Update docs, requirements, changelog
This contains multiple commits. Each commit is indicated with a
'-'

- [ganetimgr1.7] Merge & changes for custom registration

Changes have been made since to the registration process. We keep
the newer version as this is a feature

upgrade to django 1.7.7

- Completely remove South dependencies

Django v1.7 comes with built-in support for migrations. This patch drops
the South dependencies from the project in favor of the new
functionality.

- contrib: add a Gunicorn example for Debian Jessie

... and rename the previous config to ganetimgr_wheezy.

- Update the install.rst with the latest changes

This patch update the install page with the installation instructions
for a Debian Jessie host.

- version

- [ganetimgr1.8] Update requirements

- [ganetimgr1.8] Update Changelog