Add migrations for custom permissions
This feature needs migrations, to add new 'auth.Permission' objects in
the database. Here lies a problem: south does not detect custom permissions
in a model's 'Meta' class. 'schemamigration' does not produce migration
code.

We had to create a custom migration, to create new objects, and a backwards
one to delete them (when migrating backwards).

Reference:
http://stackoverflow.com/questions/1742021/adding-new-custom-permissions-in-django