Fix destroy bug by changing variable name
Summary:
Case:
This occured when a user tried to destoy his VM. If he clicked
on the destroy button, the application would email him a link which he
had to use to destroy his vm. The bug would occur if the user pressed
the button, reloaded the page and then pressed it again.

Details:
When issuing a destroy request, the application creates an InstanceAction
object. When resending the request, it filters the already existing requests
and invalidates them. After the loop had ended, the for loop variable "action"
shadowed (redefined) the function argument bearing the same name.
So when we tried later on to create a new action, we would pass an argument
of type "InstanceAction" to the create() method.

Test Plan: Deploy in ganetimgrvm

Reviewers: alexk

Differential Revision: https://phab.noc.grnet.gr/D1150