[ClustPersist] Fix disabled cluster still appears
"InstanceApplicationReviewForm"'s field "cluster" included all the
clusters that had disabled=False and disable_instance_creation=False.
This was stored in the Field's "choices" attribute. This though was
populated on project startup so if anyone made a change in realtime
it would not be visible until server was reloaded.

Reference:
http://stackoverflow.com/questions/5586803/cache-problem-with-django-forms

Fix:
Populating the "cluster" field is now done on form's "init" method.