[IndexVideo] Fix bug with modal staying on top
Previously the video modal would always be on top of the page but hidden.
This created a huge problem, as you could not click anything below the modal.
This commit fixes that issue. On page load the modal has display:none; and
whenever the trigger is clicked (actually, whenever the modal is shown / hidden)
it's visibility is toggled.