[ganetimgr1.8] Remove `null=True` from `ManyToManyField`
We kept getting the following:

* apply.Organization.users: (fields.W340) null has no effect on ManyToManyField
* ganeti.Network.groups: (fields.W340) null has no effect on ManyToManyField

Removed `null=True` as Django 1.8 now suggests.
This produced migrations too, in apps `ganeti` and `apply`. Especially `ganeti` app
had no initial migrations so they were created too. If this operation (changing this
field) produced some actual SQL code we would have to create the initial migrations
and then a second for this operation, but this is not the case since removing `null
= True` does nothing in SQL level.

Interesting issue:

When trying to test migrations for app `ganeti`, e.g. added a new field in a model
and ran `./manage.py makemigrations` we got an error. The issue was that in
`apply/admin.py` we referenced models from `ganeti` app in module-import time.

This is a problem in Django > 1.7.

Fix: Removed the relevant code (the one populating the admin ListView actions)
in `apply/admin.py` since it was no longer relevant. Issue fixed.