Ganetimgr 1.4 - 1.8 migration
Summary:
[ganetimgr1.7] Start merging & resolving conflicts

Starting the upgrade to Django 1.7. As a lot of time has passed since
this was implemented, we expect to have various conflicts.

upgrade to django 1.5

[ganetimgr1.7] commit merge

upgrade to django 1.6

change mimetype keyword to content_type in HttpResponse kwargs

[ganetimgr1.7] Merge & changes for custom registration

Changes have been made since to the registration process. We keep
the newer version as this is a feature

upgrade to django 1.7.7

Update jsonfield module to 'v1.0.3' version

Currently, the 'ganeti' app uses the django-jsonfield module in order to
store JSON fields to its models. This patch, updates this module to its
latest upstream version 1.0.3, that contains various bug fixes and also
adds support for Django v1.8.

For more details, see: https://github.com/bradjasper/django-jsonfield

Add the ALLOWED_HOST option in settings.py.dist

Since Django v1.5, the ALLOWED_HOSTS option is required in production.

The new ALLOWED_HOSTS setting validates the request's Host header and
protects against host-poisoning attacks. This setting is now required
whenever DEBUG is set to False. For more details see:

https://docs.djangoproject.com/en/1.8/ref/settings/#allowed-hosts

Remove the deprecated 'setup_environ' method call

This patch removes the call to setup_environ() method from the watcher
module, that has been removed from the django.core.management module in
Django v1.6.

Completely remove South dependencies

Django v1.7 comes with built-in support for migrations. This patch drops
the South dependencies from the project in favor of the new
functionality.

Use json module everywhere in the project

Django v1.5 deprecates 'django.utils.simplejson' in favor of the
built-in python's json module.

This patch updates the project's code to use the json module everywhere.

Refactor the action function in 'apply' admin model

This patch refactors the action method of the 'apply/admin.py' module in
order to improve the code's readability without affecting the core
functionality.

In details, we subtract the action helper method '_submit_applications'
from the 'make_fast_create_actions' method in a separate method named
'make_submit_applications'. This method new is called using the
'functools.partial()' method. Finally, we also renamed the
'ApplicationAdmin' model to 'InstanceApplicationAdmin' since it refers
to the InstanceApplication model.

add test settings

Replace gevent.pool.imap with its map equivalent

In gevent v1.0, the implementation of the 'gevent.pool.imap' has been
modified to return an iterator as its result. In ganetimgr, this
modification results in no data shown to the user since this iterator
is never consumed by the application.

To fix this issue, we replace the call to imap with the map equivalent,
which makes an internal call the to imap method and then consumes it
too. Finally, we also removed the call to 'gevent.pool.join' method
since it is not necessary anymore.

test settings and settings.dist for django 1.7

contrib: add a Gunicorn example for Debian Jessie

... and rename the previous config to ganetimgr_wheezy.

Update the install.rst with the latest changes

This patch update the install page with the installation instructions
for a Debian Jessie host.

fix reverse password change

version

[ganetimgr1.7] Fix problematic migrations

The migrations were messed up, possibly because since the Django 1.7 branch
was introduced more migrations have been added.

The problem was quite cryptic, kept getting the following error message:
django.core.exceptions.ImproperlyConfigured:
	Application labels aren't unique, duplicates: apply

Searches kept pointing to ways to actually rename your app if it has the same
label with another one, as Django 1.7 no longer allows this. This was not useful
though, as no duplicate app existed. Not really sure what the problem actually was,
but deleting existing migrations in the app fixed it.

Afterwards, though, we could not create migrations for certain apps (eg. notifications).
This happened as notifications had already some south migrations, probably trash left while
merging the commits from 1.7 branch.

To fix this:

1) All (django) migrations were deleted
2) All south-migrations for every app were added to <app>/south_migrations directory.
3) ./manage.py makemigrations to create new initial django migrations
4) ./manage.py migrate to (fake) apply the new migrations.

[ganetimgr1.7] Add correct view for registration patch

[ganetimgr1.7] Correct language url in base template

[ganetimgr1.7] Add missing urls file

Temporary changes in registration (not working)

[ganetimgr1.7] Replace deprecated close_connection() with close_old_connections()

Fix registration profile issues

Remove old registration

[ganetimgr1.7] Remove deprecated `ValidationError` imports

`ValidationError` used to be imported from `forms.utils` or
`forms.util`. This is deprecated. Import it from `django.core.exceptions`.

[ganetimgr1.7] Add correct registration urls

Registration now should follow admin activation backend. Change
imported urls to use that.

[ganetimgr1.8] Refactor registration templates

* Change templates names to the ones django-registration-redux
    expects, so that our templates are used instead of the default
    ones
* Changes in content (more suitable text / headings etc)

[ganetimgr1.8] Changes in translations

* New translations
* Changes in some old ones

[ganetimgr1.8] Remove unused code, PEP8 fixes

[ganetimgr1.8] Add custom `RegistrationView`

When deprecating our old `GanetimgrBackend` for user registration,
an issue occured. Fields `User.first_name`, `User.last_name` were
assigned manually to the `User` object. Also, a `UserProfile` object
was created manually and linked to the `User`.

To implement these in a clean manner, we have overriden the
`RegistrationView`.`register` function. Our version calls the default
function, performs some extra functionality and then returns the created
user (to conform with how the default `register` method behaved).

To do so, we have changed the '^register/' url mapping, to use our
view instead of the default one.

Changes:

* url mappings
* delete unused code
* implement custom registration functionality (view)

[ganetimgr1.8] Remove `null=True` from `ManyToManyField`

We kept getting the following:

* apply.Organization.users: (fields.W340) null has no effect on ManyToManyField
* ganeti.Network.groups: (fields.W340) null has no effect on ManyToManyField

Removed `null=True` as Django 1.8 now suggests.
This produced migrations too, in apps `ganeti` and `apply`. Especially `ganeti` app
had no initial migrations so they were created too. If this operation (changing this
field) produced some actual SQL code we would have to create the initial migrations
and then a second for this operation, but this is not the case since removing `null
= True` does nothing in SQL level.

Interesting issue:

When trying to test migrations for app `ganeti`, e.g. added a new field in a model
and ran `./manage.py makemigrations` we got an error. The issue was that in
`apply/admin.py` we referenced models from `ganeti` app in module-import time.

This is a problem in Django > 1.7.

Fix: Removed the relevant code (the one populating the admin ListView actions)
in `apply/admin.py` since it was no longer relevant. Issue fixed.

[ganetimgr1.8] Remove unused `regbackends`

Test Plan: staging

Reviewers: #developers, alexk

Differential Revision: https://phab.noc.grnet.gr/D1922