[netwname] Fix network name in instance creation
Summary:
When sending an instance creation job, the back end would send,
among others, the pair

```
'network': <value>
```

where <value>=`InstanceApplication.instanceparams['network']`.
This field was assigned when reviewing an application (or submitting
one as admin), and contained the `value` of the network selected
from the "Network" `select2` (in the web UI). This `value` was
created using the `link` of every network. This is a legacy convention,
since `link` is no longer used to determine the network logic-wise.
Thus, remove this logic and instead of `link` use `network` which
is actually the network name, as returned from `ganeti`. This,
then, propagates to the value used in instance creation, which
is our goal.

Test Plan: netdev2

Reviewers: alexk, kargig

Reviewed By: kargig

Differential Revision: https://phab.noc.grnet.gr/D2114