Fix policy mem/speed issues with new data struct
- Creates new class called MiniInstance to store details instead of the
full Instance objects. Memory footprint for Instance objects is really
big and we only need users/cluster/groups/tags/name properties
- Creates new data structure (dict) to provide faster user-vm lookup.
Through an iteration of all instances we create a dict with usernames as
keys and list of MiniInstance objects to allow faster lookup when
searching which vms are affected by policy. Speed optimization with this
change is >> 200%