Refactor Instance object initialization
- Removes list-users/organizations/groups/applications/networks from the
initialization parameters of the Instance object. There is no need to
pass these data as objects since these data can be found through the
preloading of the extra info. The preloading happened before but instead
of using the data to filter through and find the needed info, every
object (user/group etc) was passed to each Instance object. This
resulted in big (memory-wise) footprint for each Instance object
- Adds cached_data (default None) parameter in initialization of the
Instance object to allow the constructor to filter through the stuff it
needs (through the tag use)
- Adds a tag prefix resolver function to minimize the if/else statements
during the update of the Instance object