Cache all instances data for quick notif lookups
- Adds a new cache key (all:instances) to store all the instance data
(json) to allow quicker searching when looking up an instance to send a
notification email. The value is a dict {instance-name: instance-data}.
This way we can avoid using the all/filter functions of Instance
which required every cluster to initialize all Instance objects and then
search. We now use the json data to search and then craft a partial
Instance object (without cluster reference since we don't really need it
when we send a notification) to use for looking up users/groups.
- Solves tickets VIMA-1375 and VIMA-1376