Provide generic support for OS images
Summary:
- Adds generic support for OS images through meta files. An IMAGES_URL
now contains a list of meta files (marked with IMG_META_SFX suffix)
which contain all the information about where to find and how to
configure (serve to ganeti) the supported OS images. The meta files
contain the json structed information that will be passed to ganeti
- Adds support to have independent url for the images and the meta
files. Under `osparams` key in the meta file, the `img_id` key can be
set to serve the image under a different url.
- Checks that img_id exists. We can't really validate that the img_id is correct
because every provider behaves differently. So we simply check that img_id
is present and if not we filter the image out

Reviewers: #developers, #servers, alexaf

Reviewed By: #servers, alexaf

Subscribers: ivan, alexaf, #servers, #developers

Differential Revision: https://phab.noc.grnet.gr/D4491